feat(user-interviews): move api+webhooks into presentation/, enforce isolation
Final step of isolating the user_interviews product behind a facade.

- Moves backend/api.py to backend/presentation/views.py (git rename, so
  blame is preserved).
- Moves backend/webhooks.py to backend/presentation/webhooks.py.
- Updates posthog/api/__init__.py and posthog/urls.py to import from
  the new presentation paths.
- Updates internal test patch paths to point at the new module
  locations.
- Adds user_interviews to the canonical isolated-products interface
  block in tach.toml and registers a product-specific block exposing
  presentation.webhooks (used by posthog/urls.py for the Vapi handlers).
- Adds backend:contract-check to package.json so turbo-discover treats
  the product as isolated for selective CI.
- Adds the two existing presentation->models imports to the
  import-linter allowlist as TODOs; future refactors will route them
  through the facade.

`tach check --interfaces` and `lint-imports` both pass cleanly with
these changes.

Generated-By: PostHog Code
Task-Id: 4017f272-a0ba-4663-8e8f-8593fad729e2
